requirements.txt
evidently>=0.4.0
fastapi>=0.100.0
gradio>=4.0.0
langchain>=0.2.0
lightgbm>=4.0.0
lime>=0.2.0
matplotlib>=3.8.0
mlflow>=2.0.0
numpy>=1.26.0
openai>=1.0.0
optuna>=3.0.0
pandas>=2.0.0
plotly>=5.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
requests>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
seaborn>=0.13.0
shap>=0.45.0
torch>=2.0.0
transformers>=4.0.0
uvicorn>=0.20.0
xgboost>=2.0.0
